Rename BaseFrontmatterSchema -> PageFrontmatterSchema
