+++ apps/site/src/app/employee-privacy-policy/page.tsx
@@ -4,7 +4,7 @@ import employeePrivacyPolicyMarkdown from '@/content/pages/employee-privacy-poli
 
 import { createMetadata } from '@/utils/createMetadata'
 
-import { MarkdownPageSchema } from '@/schemas/FrontmatterSchema'
+import { MarkdownPageSchema } from '@/schemas/PageFrontmatterSchema'
 
 import { MarkdownPage } from '@/components/MarkdownPage'
