+++ apps/site/src/app/filecoin-plus/page.tsx
@@ -9,7 +9,7 @@ import { graphicsData } from '@/data/graphicsData'
 import { createMetadata } from '@/utils/createMetadata'
 import { extractDomain } from '@/utils/extractDomain'
 
-import { BaseFrontmatterSchema } from '@/schemas/FrontmatterSchema'
+import { PageFrontmatterSchema } from '@/schemas/PageFrontmatterSchema'
 
 import { Badge } from '@/components/Badge'
 import { BadgeCardGrid } from '@/components/BadgeCardGrid'
@@ -28,9 +28,9 @@ import { aboutData } from './data/aboutData'
 import { applicationData } from './data/applicationData'
 import { generateStructuredData } from './utils/generateStructuredData'
 
-const { header, seo } = BaseFrontmatterSchema.parse(attributes)
+const { header, seo } = PageFrontmatterSchema.parse(attributes)
 const { header: allocatorsHeader } =
-  BaseFrontmatterSchema.parse(allocatorsAttributes)
+  PageFrontmatterSchema.parse(allocatorsAttributes)
 
 export const metadata = createMetadata({
   seo: {
