+++ apps/site/src/app/grants/schemas/FrontmatterSchema.ts
@@ -1,8 +1,8 @@
 import { z } from 'zod'
 
-import { BaseFrontmatterSchema } from '@/schemas/FrontmatterSchema'
 import { MarkdownEntryPathSchema } from '@/schemas/MarkdownEntryPathSchema'
+import { PageFrontmatterSchema } from '@/schemas/PageFrontmatterSchema'
 
-export const FrontmatterSchema = BaseFrontmatterSchema.extend({
+export const FrontmatterSchema = PageFrontmatterSchema.extend({
   featured_grant_graduates: z.array(MarkdownEntryPathSchema),
 })
