+++ apps/site/src/app/privacy-policy/page.tsx
@@ -4,7 +4,7 @@ import privacyPolicyMarkdown from '@/content/pages/privacy-policy.md'
 
 import { createMetadata } from '@/utils/createMetadata'
 
-import { MarkdownPageSchema } from '@/schemas/FrontmatterSchema'
+import { MarkdownPageSchema } from '@/schemas/PageFrontmatterSchema'
 
 import { MarkdownPage } from '@/components/MarkdownPage'
